requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-WTF==0.15.1
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mysql-connector-python==8.0.25
numpy==1.20.1
opencv-python==4.4.0.46
pandas==1.2.5
Pillow==8.0.1
protobuf==3.17.3
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
